+++ datatest/src/commonMain/kotlin/at/asitplus/test/Terminal.vararg.kt
@@ -35,6 +35,37 @@ fun <Data> TestSuiteScope.withData(
     action
 )
 
+
+/**
+ * Executes a test for each provided data parameter.
+ *
+ * @param parameters The data parameters to test with
+ * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
+ * @param maxLength maximum length of test element name (not display name)
+ * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
+ * @param testConfig Optional test configuration
+ * @param action Test action to execute for each parameter
+ */
+fun <Data> TestSuiteScope.withData(
+    vararg parameters: Pair<String,Data>,
+    compact: Boolean = DataTest.compactByDefault,
+    maxLength: Int = DataTest.defaultTestNameMaxLength!!,
+    displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength!!,
+    prefix: String = "",
+    testConfig: TestConfig = TestConfig,
+    action: suspend Test.ExecutionScope.(Data) -> Unit
+) = withDataInternal(
+    parameters.asSequence(),
+    testConfig,
+    compact,
+    maxLength,
+    displayNameMaxLength,
+    prefix,
+    action
+)
+
+
 /**
  * Executes a test for each provided data parameter.
  * Uses provided function to generate test names.